adding first test
